% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emirror.R
\name{emirror}
\alias{emirror}
\title{emirror}
\usage{
emirror(top, bottom, line, log10 = TRUE, yaxis, opacity = 1,
  toptitle = NULL, bottomtitle = NULL, annotate_var, annotate_p,
  highlight_var, highlight_p, highlighter = "red", color1 = "#AAAAAA",
  color2 = "#4D4D4D", groupcolors, rotatelabels = FALSE, labelangle,
  file = "emirror", relhgts = c(1/2, 1/2), hgt = 7, wi = 12,
  res = 300)
}
\arguments{
\item{top}{data frame, columns one and two must be Variable and pvalue; Group, Shape and Color optional}

\item{bottom}{data frame, columns one and two must be Variable and pvalue; Group, Shape and Color optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{toptitle}{optional string for plot title}

\item{bottomtitle}{optional string for plot title}

\item{annotate_var}{list of variables to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{labelangle}{angle to rotate}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}

\item{rotatelabel}{boolean, rotate axis labels?}

\item{relghts}{relative heights for the plots, useful if using long axis labels}
}
\value{
png image(s)
}
\description{
Create mirrored Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
Suggested: RColorBrewer, ggrepel
}
\examples{
emirror(top, bottom, line, file="emirror", hgt=7, wi=12, res=300 )
}
