% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewasb-data.R
\docType{data}
\name{ewas.b}
\alias{ewas.b}
\title{Example EWAS data}
\format{
An object of class data.frame
}
\usage{
data(ewas.b)
}
\description{
Example EWAS data
}
\examples{
data(ewas.b)
}
\keyword{datasets}
