% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewast-data.R
\docType{data}
\name{ewas.t}
\alias{ewas.t}
\title{Example EWAS data}
\format{
An object of class data.frame
}
\usage{
data(ewas.t)
}
\description{
Example EWAS data
}
\examples{
data(ewas.t)
}
\keyword{datasets}
