% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasb-data.R
\docType{data}
\name{gwas.b}
\alias{gwas.b}
\title{Example GWAS data}
\format{
An object of class data.frame
}
\usage{
data(gwas.b)
}
\description{
Example GWAS data
}
\examples{
data(gwas.b)
}
\keyword{datasets}
