% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwast-data.R
\docType{data}
\name{gwas.t}
\alias{gwas.t}
\title{Example GWAS data}
\format{
An object of class data.frame
}
\usage{
data(gwas.t)
}
\description{
Example GWAS data
}
\examples{
data(gwas.t)
}
\keyword{datasets}
