% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemirror.R
\name{iemirror}
\alias{iemirror}
\title{iemirror}
\usage{
iemirror(
  top,
  bottom,
  tline,
  bline,
  log10 = TRUE,
  yaxis,
  opacity = 1,
  annotate_snp,
  annotate_p,
  highlight_snp,
  highlight_p,
  highlighter = "red",
  title = NULL,
  color1 = "#AAAAAA",
  color2 = "#4D4D4D",
  blockmin = -1,
  blockmax = 1,
  groupcolors,
  rotatelabels = FALSE,
  labelangle,
  freey = FALSE,
  ymax,
  ymin,
  background = "variegated",
  grpblocks = FALSE,
  file = "iemirror",
  hgtratio = 0.5,
  hgt = 7,
  wi = 12,
  res = 300
)
}
\arguments{
\item{top}{data frame, columns one and two must be Variable, pvalue, and Group; Shape, Color, Hover, Link optional
where 'Hover' contains tooltip information overriding the default (Variable, Group)
and 'Link' contains a search query, ex. "https://en.wikipedia.org/wiki/"}

\item{bottom}{data frame, columns one and two must be Variable, pvalue, and Group; Shape, Color, Hover, Link optional
where 'Hover' contains tooltip information overriding the default (Variable, Group)
and 'Link' contains a search query, ex. "https://en.wikipedia.org/wiki/"}

\item{tline}{list of pvalues to draw red threshold lines in top plot}

\item{bline}{list of pvalues to draw red threshold lines in bottom plot}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis in the format c("top", "bottom"), automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{annotate_p}{list of pvalue thresholds to annotate in the order of c(p_top, p_bottom)}

\item{highlight_p}{list of pvalue thresholds to highlight in the order of c(p_top, p_bottom)}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named vector of colors where names correspond to data in 'Color' column}

\item{rotatelabels}{boolean, rotate axis labels?}

\item{labelangle}{angle to rotate}

\item{freey}{allow y-axes to scale with data}

\item{background}{variegated or white}

\item{grpblocks}{boolean, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{hgtratio}{height ratio of plots, equal to top plot proportion}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}

\item{toptitle}{optional string for plot title}

\item{bottomtitle}{optional string for plot title}

\item{annotate_var}{vector of variables to annotate}

\item{highlight_var}{vector of variables to highlight}

\item{type}{plot type/extension}
}
\value{
png image
}
\description{
Create interactive mirrored Manhattan plots for EWAS
Dependencies: ggplot2, ggiraph
Suggested: ggrepel
}
\examples{
data(ewas.t)
data(ewas.b)
emirror(top=ewas.t, bottom=ewas.b, annotate_p = c(0.0001, 0.0005), 
highlight_p=c(0.0001, 0.0005), highlighter="green", toptitle = "EWAS Comparison Example: Data 1", 
bottomtitle = "EWAS Comparison Example: Data 2")
}
