% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igmirror.R
\name{igmirror}
\alias{igmirror}
\title{igmirror}
\usage{
igmirror(
  top,
  bottom,
  tline,
  bline,
  chroms = c(1:22, "X", "Y"),
  log10 = TRUE,
  yaxis,
  opacity = 1,
  highlight_snp,
  highlight_p,
  highlighter = "red",
  title = NULL,
  chrcolor1 = "#AAAAAA",
  chrcolor2 = "#4D4D4D",
  chrblockmin = -1,
  chrblockmax = 1,
  freey = FALSE,
  ymax,
  ymin,
  background = "variegated",
  chrblocks = FALSE,
  file = "igmirror",
  hgt = 7,
  wi = 12,
  res = 300
)
}
\arguments{
\item{top}{data frame, must contain SNP, CHR, POS, pvalue columns, optional Shape, Hover, Link
where 'Hover' contains tooltip information overriding the default (SNP, chr:position)
and 'Link' contains a search query, ex. "https://www.ncbi.nlm.nih.gov/snp/rs944173"}

\item{bottom}{data frame, must contain SNP, CHR, POS, pvalue, columns, optional Shape, Hover, Link
where 'Hover' contains tooltip information overriding the default (SNP, chr:position)
and 'Link' contains a search query, ex. "https://www.ncbi.nlm.nih.gov/snp/rs944173"}

\item{tline}{list of pvalues to draw red threshold lines in top plot}

\item{bline}{list of pvalues to draw red threshold lines in bottom plot}

\item{chroms}{list of chromosomes to plot in the order desired, default c(1:22, "X", "Y")}

\item{log10}{plot -log10() of pvalue column, logical}

\item{yaxis}{title for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0-1, useful for dense plots}

\item{highlight_snp}{vector of snps to highlight}

\item{highlight_p}{list of pvalue thresholds to annotate in the order of c(p_top, p_bottom)}

\item{highlighter}{color to highlight}

\item{title}{figure title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{chrblockmin}{minimum y value for chromosome block}

\item{chrblockmax}{maximum y value for chromosome block}

\item{freey}{allow y-axes to scale with the data}

\item{ymax}{override the auto set y axis limit}

\item{ymin}{override the auto set y axis limit}

\item{background}{variegated or white}

\item{chrblocks}{logical, turns on x-axis chromosome marker blocks}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create interactive mirrored Manhattan plots for GWAS
Dependencies: ggplot2, ggiraph
Suggested: ggrepel
}
\examples{
data(gwas.t)
data(gwas.b)
gwas.t$Link <- paste0("https://www.ncbi.nlm.nih.gov/snp/", gwas.t$SNP)
gwas.b$Link <- paste0("https://www.ncbi.nlm.nih.gov/snp/", gwas.b$SNP)
igmirror(gwas.t, gwas.b)
}
