% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phemirror.R
\name{phemirror}
\alias{phemirror}
\title{phemirror}
\usage{
phemirror(
  top,
  bottom,
  phegroup,
  tline,
  bline,
  chroms = c(1:22, "X", "Y"),
  log10 = TRUE,
  yaxis,
  opacity = 1,
  annotate_snp,
  annotate_p,
  highlight_snp,
  highlight_p,
  highlighter = "red",
  toptitle = NULL,
  bottomtitle = NULL,
  chrcolor1 = "#AAAAAA",
  chrcolor2 = "#4D4D4D",
  groupcolors,
  freey = FALSE,
  background = "variegated",
  chrblocks = TRUE,
  file = "phemirror",
  hgtratio = 0.5,
  hgt = 7,
  wi = 12,
  res = 300
)
}
\arguments{
\item{top}{data frame, if not plato or plink format, must contain PHE, SNP, CHR, POS, pvalue, columns, optional Shape}

\item{bottom}{data frame, if not plato or plink format, must contain PHE, SNP, CHR, POS, pvalue, columns, optional Shape}

\item{phegroup}{optional grouping file for phenotypes, must contain PHE and Group columns}

\item{tline}{list of pvalues to draw red threshold lines in top plot}

\item{bline}{list of pvalues to draw red threshold lines in bottom plot}

\item{chroms}{list of chromosomes to plot in the order desired, default c(1:22, "X", "Y")}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis in the format c("top", "bottom"), automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0-1, useful for dense plots}

\item{annotate_snp}{vector of RSIDs to annotate}

\item{annotate_p}{list of pvalue thresholds to annotate in the order of c(p_top, p_bottom)}

\item{highlight_snp}{vector of snps to highlight}

\item{highlight_p}{list of pvalue thresholds to annotate in the order of c(p_top, p_bottom)}

\item{highlighter}{color to highlight}

\item{toptitle}{optional string for top plot title}

\item{bottomtitle}{optional string for bottom plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named vector of colors where names correspond to data in 'PHE' or 'Group' column}

\item{freey}{allow y-axes to scale with the data}

\item{background}{variegated or white}

\item{chrblocks}{boolean, turns on x-axis chromosome marker blocks}

\item{file}{file name of saved image}

\item{hgtratio}{height ratio of plots, equal to top plot proportion}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create mirrored Manhattan plots for PheWAS
Dependencies: ggplot2, gridExtra, RColorBrewer
Suggested: ggrepel
}
\examples{
data(phewas.t)
data(phewas.b)
phemirror(top=phewas.t, bottom = phewas.b, 
toptitle = "PheWAS Example: Data 1", bottomtitle = "PheWAS Example: Data 2")
}
