% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewasb-data.R
\docType{data}
\name{phewas.b}
\alias{phewas.b}
\title{Example PheWAS data}
\format{
An object of class data.frame
}
\usage{
data(phewas.b)
}
\description{
Example PheWAS data
}
\examples{
data(phewas.b)
}
\keyword{datasets}
