% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewast-data.R
\docType{data}
\name{phewas.t}
\alias{phewas.t}
\title{Example PheWAS data}
\format{
An object of class data.frame
}
\usage{
data(phewas.t)
}
\description{
Example PheWAS data
}
\examples{
data(phewas.t)
}
\keyword{datasets}
