% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqmirror.R
\name{qqmirror}
\alias{qqmirror}
\title{qqmirror}
\usage{
qqmirror(top, bottom, CI = 0.95, opacity = 1, groupcolors,
  splittop = NULL, splitbottom = NULL, highlight_p, highlight_name,
  annotate_p, annotate_name, highlighter = "red", tline, bline, background,
  toptitle = NULL, bottomtitle = NULL, file = "qqmirror", wi = 6,
  hgt = 12, hgtratio = 0.5, res = 300)
}
\arguments{
\item{top}{dataframe with at least one column, p-value; Color, Shape, and Name optional}

\item{bottom}{dataframe with at least one column, p-value; Color, Shape, and Name optional}

\item{CI}{two-sided confidence interval, default 0.95}

\item{opacity}{point opacity, default 1}

\item{groupcolors}{named list of colors corresponding to data in Group column}

\item{splittop}{if data contains Color and/or Shape, indicate variable(s) by which the data should be subsetted for calculating CIs}

\item{splitbottom}{if data contains Color and/or Shape, indicate variable(s) by which the data should be subsetted for calculating CIs}

\item{highlight_p}{p-value threshold to highlight}

\item{highlight_name}{list of names to highlight, dataframe must include a Name column}

\item{annotate_p}{p-value threshold to annotate, dataframe must include a Name column}

\item{annotate_name}{list of names to annotate, dataframe must include a Name column}

\item{highlighter}{highlighter color}

\item{background}{can change to "white"}

\item{toptitle}{title for top plot}

\item{bottomtitle}{title for bottom plot}

\item{file}{filename}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{hgtratio}{height ratio for plots}

\item{res}{resolution of plot}

\item{line}{draw a red line at pvalue threshold (observed)}
}
\value{
png image
}
\description{
Create qqplots with an assumed uniform distribution
}
\examples{
data(gwas)
qqdat <- data.frame(pvalue=gwas$pvalue, Color=gwas$Frame)
qqmirror(top=qqdat, bottom=qqdat, opacity=0.6, splittop="Color", splitbottom="Color", toptitle="Example: Top", bottomtitle="Example: Bottom")
}
