% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqunif.R
\name{qqunif}
\alias{qqunif}
\title{qqunif}
\usage{
qqunif(
  d,
  CI = 0.95,
  opacity = 1,
  groupcolors,
  splitby = NULL,
  highlight_p,
  highlight_name,
  annotate_p,
  annotate_name,
  highlighter = "red",
  line,
  slope = 1,
  background,
  title
)
}
\arguments{
\item{d}{dataframe with at least one column, p-value; Color, Shape, and Name optional}

\item{CI}{two-sided confidence interval, default 0.95}

\item{opacity}{point opacity, default 1}

\item{groupcolors}{named vector of colors corresponding to data in Group column}

\item{splitby}{if data contains Color and/or Shape, indicate variable(s) by which the data should be subsetted for calculating CIs}

\item{highlight_p}{p-value threshold to highlight}

\item{highlight_name}{vector of names to highlight, dataframe must include a Name column}

\item{annotate_p}{p-value threshold to annotate, dataframe must include a Name column}

\item{annotate_name}{vector of names to annotate, dataframe must include a Name column}

\item{highlighter}{highlighter color}

\item{line}{draw a red line at pvalue threshold (observed)}

\item{slope}{abline slope}

\item{background}{can change to "white"}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
Create qqplots with an assumed uniform distribution
}
