% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emirror.R
\name{emirror}
\alias{emirror}
\title{emirror}
\usage{
emirror(top, bottom, tline, bline, log10 = TRUE, yaxis, opacity = 1,
  toptitle = NULL, bottomtitle = NULL, annotate_var, annotate_p,
  highlight_var, highlight_p, highlighter = "red", color1 = "#AAAAAA",
  color2 = "#4D4D4D", groupcolors, rotatelabels = FALSE, labelangle,
  freey = FALSE, background = "variegated", grpblocks = FALSE,
  file = "emirror", hgtratio = 0.5, hgt = 7, wi = 12, res = 300)
}
\arguments{
\item{top}{data frame, columns one and two must be Variable, pvalue, and Group; Shape and Color optional}

\item{bottom}{data frame, columns one and two must be Variable, pvalue, and Group; Shape and Color optional}

\item{tline}{optional pvalue threshold to draw red line at in top plot}

\item{bline}{optional pvalue threshold to draw red line at in bottom plot}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis in the format c("top", "bottom"), automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{toptitle}{optional string for plot title}

\item{bottomtitle}{optional string for plot title}

\item{annotate_var}{list of variables to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named list of colors where names correspond to data in 'Color' column}

\item{labelangle}{angle to rotate}

\item{freey}{allow y-axes to scale with data}

\item{background}{variegated or white}

\item{grpblocks}{boolean, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{hgtratio}{height ratio of plots, equal to top plot proportion}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}

\item{rotatelabel}{boolean, rotate axis labels?}
}
\value{
png image(s)
}
\description{
Create mirrored Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
Suggested: RColorBrewer, ggrepel
}
\examples{
emirror(top, bottom, line, file="emirror", hgt=7, wi=12, res=300 )
}
