% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmirror.R
\name{gmirror}
\alias{gmirror}
\title{gmirror}
\usage{
gmirror(top, bottom, tline, bline, log10 = TRUE, yaxis, opacity = 1,
  annotate_snp, annotate_p, toptitle = NULL, bottomtitle = NULL,
  highlight_snp, highlight_p, highlighter = "red", chrcolor1 = "#AAAAAA",
  chrcolor2 = "#4D4D4D", file = "gmirror", hgt = 7, wi = 12,
  res = 300)
}
\arguments{
\item{top}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, optional Shape}

\item{bottom}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, optional Shape}

\item{tline}{optional pvalue threshold to draw red line at in top plot}

\item{bline}{optional pvalue threshold to draw red line at in bottom plot}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{annotate_snp}{list of RSIDs to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{toptitle}{optional string for top plot title}

\item{bottomtitle}{optional string for bottom plot title}

\item{highlight_snp}{list of snps to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create mirrored Manhattan plots for GWAS
Dependencies: ggplot2, gridExtra
Suggested: RColorBrewer, ggrepel
}
\examples{
gmirror(top, bottom, line, log10, yaxis, opacity, annotate_snp, annotate_p, title, chrcolor1, chrcolor2, file, hgt, wi, res)
}
