% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmirror.R
\name{gmirror}
\alias{gmirror}
\title{gmirror}
\usage{
gmirror(
  top,
  bottom,
  tline,
  bline,
  log10 = TRUE,
  yaxis,
  opacity = 1,
  annotate_snp,
  annotate_p,
  toptitle = NULL,
  bottomtitle = NULL,
  highlight_snp,
  highlight_p,
  highlighter = "red",
  chrcolor1 = "#AAAAAA",
  chrcolor2 = "#4D4D4D",
  freey = FALSE,
  background = "variegated",
  chrblocks = FALSE,
  file = "gmirror",
  hgt = 7,
  hgtratio = 0.5,
  wi = 12,
  res = 300
)
}
\arguments{
\item{top}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, optional Shape}

\item{bottom}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, optional Shape}

\item{tline}{optional pvalue threshold to draw red line at in top plot}

\item{bline}{optional pvalue threshold to draw red line at in bottom plot}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis in the format c("top", "bottom"), automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{annotate_snp}{vector of RSIDs to annotate}

\item{annotate_p}{list of pvalue thresholds to annotate in the order of c(p_top, p_bottom)}

\item{toptitle}{optional string for top plot title}

\item{bottomtitle}{optional string for bottom plot title}

\item{highlight_snp}{vector of snps to highlight}

\item{highlight_p}{list of pvalue thresholds to highlight in the order of c(p_top, p_bottom)}

\item{highlighter}{color to highlight}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{freey}{allow y-axes to scale with the data}

\item{background}{variegated or white}

\item{chrblocks}{boolean, turns on x-axis chromosome marker blocks}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{hgtratio}{height ratio of plots, equal to top plot proportion}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create mirrored Manhattan plots for GWAS
Dependencies: ggplot2, gridExtra
Suggested: RColorBrewer, ggrepel
}
\examples{
data(gwas.t)
data(gwas.b)
gmirror(top=gwas.t, bottom=gwas.b, tline=0.05/nrow(gwas.t), bline=0.05/nrow(gwas.b), 
toptitle="GWAS Comparison Example: Data 1", bottomtitle = "GWAS Comparison Example: Data 2", 
highlight_p = c(0.05/nrow(gwas.t), 0.05/nrow(gwas.b)), highlighter="green")
}
