% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemirror.R
\name{iemirror}
\alias{iemirror}
\title{iemirror}
\usage{
iemirror(
  top,
  bottom,
  tline,
  bline,
  log10 = TRUE,
  yaxis = NULL,
  xaxis = NULL,
  levs,
  opacity = 1,
  highlight_var,
  highlight_p,
  highlighter = "red",
  title = NULL,
  color1 = "#AAAAAA",
  color2 = "#4D4D4D",
  blockmin = -1,
  blockmax = 1,
  groupcolors,
  rotatelabels = FALSE,
  labelangle,
  freey = FALSE,
  ymax,
  ymin,
  background = "variegated",
  grpblocks = FALSE,
  file = "iemirror",
  hgt = 7,
  wi = 12,
  res = 300
)
}
\arguments{
\item{top}{data frame, columns one through three must be Variable, pvalue, and Group; Shape, Color, Hover, Link optional
where 'Hover' contains tooltip information overriding the default (Variable, Group)
and 'Link' contains a search query, ex. "https://en.wikipedia.org/wiki/"}

\item{bottom}{data frame, columns one and two must be Variable, pvalue, and Group; Shape, Color, Hover, Link optional
where 'Hover' contains tooltip information overriding the default (Variable, Group)
and 'Link' contains a search query, ex. "https://en.wikipedia.org/wiki/"}

\item{tline}{list of pvalues to draw red threshold lines in top plot}

\item{bline}{list of pvalues to draw red threshold lines in bottom plot}

\item{log10}{plot -log10() of pvalue column, logical}

\item{yaxis}{title for y-axis, automatically set if log10=TRUE}

\item{xaxis}{title for x-axis}

\item{levs}{if 'Color' column give, vector containing the order the values should appear in on the legend}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{highlight_var}{vector of variables to highlight}

\item{highlight_p}{list of pvalue thresholds to highlight in the order of c(p_top, p_bottom)}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named vector of colors where names correspond to data in 'Color' column}

\item{rotatelabels}{logical, rotate axis labels?}

\item{labelangle}{angle to rotate}

\item{freey}{allow y-axes to scale with data}

\item{background}{variegated or white}

\item{grpblocks}{logical, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}

\item{toptitle}{optional string for plot title}

\item{bottomtitle}{optional string for plot title}
}
\value{
png image
}
\description{
Create interactive mirrored Manhattan plots for EWAS
Dependencies: ggplot2, ggiraph
Suggested: ggrepel
}
\examples{
data(ewas.t)
data(ewas.b)
ewas.t$Link <- paste0("https://www.google.com/search?q=", ewas.t$Variable)
ewas.b$Link <- paste0("https://www.google.com/search?q=", ewas.b$Variable)
iemirror(top=ewas.t, bottom=ewas.b)
}
